@@ -16,69 +16,72 @@
   "author": "Kinggeorge Magaga",
   "license": "Apache-2.0",
   "dependencies": {
-    "@angular-material-extensions/link-preview": "^1.0.2",
-    "@angular/animations": "^9.0.0-next.4",
-    "@angular/cdk": "^8.1.4",
-    "@angular/common": "^9.0.0-next.4",
-    "@angular/compiler": "^9.0.0-next.4",
-    "@angular/core": "^9.0.0-next.4",
-    "@angular/forms": "^9.0.0-next.4",
-    "@angular/http": "^7.2.1",
-    "@angular/material": "^8.1.4",
-    "@angular/platform-browser": "^9.0.0-next.4",
-    "@angular/platform-browser-dynamic": "^9.0.0-next.4",
-    "@angular/platform-server": "^9.0.0-next.4",
-    "@angular/router": "^9.0.0-next.4",
-    "@ng-bootstrap/ng-bootstrap": "^3.2.2",
-    "@ng-select/ng-select": "^2.11.2",
+    "@angular-material-extensions/link-preview": "^1.1.1",
+    "@angular/animations": "^10.2.1",
+    "@angular/cdk": "^8.2.3",
+    "@angular/common": "^10.2.1",
+    "@angular/compiler": "^10.2.1",
+    "@angular/core": "^10.2.1",
+    "@angular/forms": "^10.2.1",
+    "@angular/http": "^7.2.16",
+    "@angular/material": "^8.2.3",
+    "@angular/platform-browser": "^10.2.1",
+    "@angular/platform-browser-dynamic": "^10.2.1",
+    "@angular/platform-server": "^10.2.1",
+    "@angular/router": "^10.2.1",
+    "@ng-bootstrap/ng-bootstrap": "^7.0.0",
+    "@ng-select/ng-select": "^2.20.5",
     "@ngx-translate/core": "^12.1.2",
     "@ngx-translate/http-loader": "^5.0.0",
-    "@types/chart.js": "^2.7.31",
-    "angular-resizable-element": "^3.3.0",
-    "angular5-csv": "^0.2.10",
-    "angularx-social-login": "^1.2.6",
-    "bootstrap": "^4.3.1",
-    "chart.js": "^2.5.0",
+    "@types/chart.js": "^2.9.27",
+    "angular-resizable-element": "^3.3.3",
+    "angular5-csv": "^0.2.11",
+    "angularx-social-login": "^3.3.1",
+    "bootstrap": "^4.5.3",
+    "chart.js": "^2.9.4",
     "cordova-browser": "^6.0.0",
-    "cordova-plugin-device": "^2.0.2",
-    "core-js": "^2.6.2",
+    "cordova-plugin-device": "^2.0.3",
+    "core-js": "^2.6.11",
     "enhanced-resolve": "^3.4.1",
     "font-awesome": "^4.7.0",
     "hammerjs": "^2.0.8",
     "intersection-observer": "^0.11.0",
     "lodash": "^4.17.20",
     "ng-bootstrap": "^1.6.3",
-    "ng-in-viewport": "^6.1.3",
-    "ngx-image-compress": "^8.0.3",
-    "ngx-infinite-scroll": "^9.0.0",
+    "ng-in-viewport": "^6.1.4",
+    "ngx-image-compress": "^8.0.4",
+    "ngx-infinite-scroll": "^9.1.0",
     "node-properties-parser": "0.0.2",
-    "rxjs": "^6.5.3",
-    "rxjs-compat": "^6.3.3",
-    "typescript": "^3.5.3",
-    "zone.js": "^0.10.2"
+    "rxjs": "^6.6.3",
+    "rxjs-compat": "^6.6.3",
+    "tslib": "^2.0.0",
+    "typescript": "^3.9.7",
+    "zone.js": "~0.10.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.803.8",
-    "@angular/cli": "^8.3.2",
-    "@angular/compiler-cli": "^9.0.0-next.4",
-    "@angular/language-service": "^9.0.0-next.4",
-    "@types/jasmine": "^3.3.7",
-    "@types/jasminewd2": "~2.0.3",
-    "@types/node": "^10.12.18",
-    "codelyzer": "^4.5.0",
+    "@angular-devkit/build-angular": "~0.1002.0",
+    "@angular/cli": "^10.2.0",
+    "@angular/compiler-cli": "^10.2.1",
+    "@angular/language-service": "^10.2.1",
+    "@angular/localize": "^10.2.1",
+    "@types/jasmine": "^3.6.0",
+    "@types/jasminewd2": "^2.0.8",
+    "@types/node": "^12.11.1",
+    "codelyzer": "^5.1.2",
+    "cordova-android": "^9.0.0",
     "cordova-plugin-whitelist": "^1.3.4",
-    "jasmine-core": "^3.3.0",
-    "jasmine-spec-reporter": "^4.2.1",
-    "karma": "^5.1.1",
-    "karma-chrome-launcher": "^2.2.0",
+    "jasmine-core": "~3.5.0",
+    "jasmine-spec-reporter": "~5.0.0",
+    "karma": "~5.0.0",
+    "karma-chrome-launcher": "~3.1.0",
     "karma-cli": "^2.0.0",
-    "karma-coverage-istanbul-reporter": "~2.0.0",
-    "karma-jasmine": "^2.0.1",
-    "karma-jasmine-html-reporter": "^1.4.0",
-    "protractor": "^7.0.0",
+    "karma-coverage-istanbul-reporter": "~3.0.2",
+    "karma-jasmine": "~4.0.0",
+    "karma-jasmine-html-reporter": "^1.5.0",
+    "protractor": "~7.0.0",
     "rxjs-tslint": "^0.1.8",
     "ts-node": "~5.0.1",
-    "tslint": "^5.12.1"
+    "tslint": "~6.1.0"
   },
   "cordova": {
     "plugins": {
