@@ -1,5 +1,6 @@
 import { Component, Input, OnInit } from '@angular/core';
 import { PageSummary } from '@app/models/page/page.model';
+import { UrlViewerService } from '@app/services/url-viewer.service';
 
 @Component({
   selector: 'app-page-card',
@@ -9,7 +10,7 @@ import { PageSummary } from '@app/models/page/page.model';
 export class PageCardComponent implements OnInit {
 
   @Input() page: PageSummary;
-  constructor() { }
+  constructor(public urlViewerService:UrlViewerService) { }
 
   ngOnInit() {
   }
