@@ -20,7 +20,7 @@ export class Profile {
     timezone:string;
     country:string; 
     city:string; 
-    location:string;    
+    location:any;    
     state:string; 
     street:string; 
     zip:string; 
