@@ -35,7 +35,7 @@ import { SimpleFeedCardWidgetComponent } from './widgets/simple-feed-card-widget
 import { WelcomePageComponent } from './components/welcome-page/welcome-page.component';
 import { WfNavbarSummaryComponent } from './components/wf-navbar-summary/wf-navbar-summary.component';
 import { WfSocialHeaderComponent } from './components/wf-social-header/wf-social-header.component';
-import { AuthService as sAuthService } from "angularx-social-login";
+/*import { AuthService as sAuthService } from "angularx-social-login";*/
 import { FaintBodyOverlayComponent } from './components/faint-body-overlay/faint-body-overlay.component';
 import { WfLinkPreviewService } from '@app/libs/wf-link-preview/services/wf-link-preview.service';
 import { FeedCommentRepliesWidgetComponent } from './widgets/feed-comment-replies-widget/feed-comment-replies-widget.component';
@@ -47,13 +47,13 @@ import { FeedImagesOverlayViewComponent } from './components/feed-images-overlay
 import { AddProfilePhotoAlertWidgetComponent } from './widgets/add-profile-photo-alert-widget/add-profile-photo-alert-widget.component';
 import { OverlayModalContainerComponent } from './modals/overlay-modal-container/overlay-modal-container.component';
 import { LoadSubmitProcessDataModalComponent } from './modals/load-submit-process-data-modal/load-submit-process-data-modal.component';
-import { OverlayLoadSubmitProcessDataWidgetComponent } from './widgets/overlay-load-submit-process-data-widget/overlay-load-submit-process-data-widget.component';
+import { OverlayLoadSubmitProcessDataWidgetComponent } from './modals/overlay-load-submit-process-data-widget/overlay-load-submit-process-data-widget.component';
 import { NavigationWatchComponent } from './components/navigation-watch/navigation-watch.component';
 import { ProcessUploadImagesProgressWidgetComponent } from './widgets/process-upload-images-progress-widget/process-upload-images-progress-widget.component';
 import { ProfileFeedImageViewWidgetComponent } from '../viewer/profile/widgets/profile-feed-image-view-widget/profile-feed-image-view-widget.component';
 import { WfLinkifyModule } from '@app/libs/wf-linkify';
 import { SharedPostWidgetComponent } from './widgets/shared-post-widget/shared-post-widget.component';
-import { OverlayPostViewWidgetComponent } from './widgets/overlay-post-view-widget/overlay-post-view-widget.component';
+import { OverlayPostViewWidgetComponent } from './modals/overlay-post-view-widget/overlay-post-view-widget.component';
 import { PostCardWidgetComponent } from './widgets/post-card-widget/post-card-widget.component';
 import { WbMessagesContentComponent } from './components/wb-messages-content/wb-messages-content.component';
 import { CreatePageModalComponent } from './modals/create-page-modal/create-page-modal.component';
@@ -65,11 +65,6 @@ import { AdTextRightColumnViewComponent } from './widgets/ads/ad-text-right-colu
 import { AdTextTopBannerSmallViewComponent } from './widgets/ads/ad-text-top-banner-small-view/ad-text-top-banner-small-view.component';
 import { RightSidebarContainerComponent } from './layouts/right-sidebar-container/right-sidebar-container.component';
 import { InPageContentModalComponent } from './widgets/in-page-content-modal/in-page-content-modal.component';
-import { VerticalBarsComponent } from './widgets/loaders/vertical-bars/vertical-bars.component';
-import { LoadingBarsComponent } from './widgets/loaders/loading-bars/loading-bars.component';
-import { LoadingCirclesComponent } from './widgets/loaders/loading-circles/loading-circles.component';
-import { FoldingCubeComponent } from './widgets/loaders/folding-cube/folding-cube.component';
-import { RippleCirclesComponent } from './widgets/loaders/ripple-circles/ripple-circles.component';
 import { MyPagesListWidgetComponent } from './widgets/pages/my-pages-list-widget/my-pages-list-widget.component';
 import { WbOverlayPostFormContainerComponent } from './components/wb-overlay-post-form-container/wb-overlay-post-form-container.component';
 import { WbPageFollowButtonComponent } from './components/buttons/wb-page-follow-button/wb-page-follow-button.component';
@@ -77,11 +72,18 @@ import { WbProfileConnectButtonComponent } from './components/buttons/wb-profile
 import { WbProfileMessageButtonComponent } from './components/buttons/wb-profile-message-button/wb-profile-message-button.component';
 import { WbPageMessageButtonComponent } from './components/buttons/wb-page-message-button/wb-page-message-button.component';
 import { WbContentNotFoundPageComponent } from './pages/wb-content-not-found-page/wb-content-not-found-page.component';
-import { WbLoadingListWidgetComponent } from './widgets/loaders/wb-loading-list-widget/wb-loading-list-widget.component';
 import { DeletePostModalComponent } from './modals/delete-post-modal/delete-post-modal.component';
 import { DeleteCommentModalComponent } from './modals/delete-comment-modal/delete-comment-modal.component';
 import { DeleteConnectionModalComponent } from './modals/delete-connection-modal/delete-connection-modal.component';
 import { TranslateModule } from '@ngx-translate/core';
+import { ShowMorePostTextWidgetComponent } from './widgets/show-more-post-text-widget/show-more-post-text-widget.component';
+import { ShowMoreTextWidgetComponent } from './widgets/show-more-text-widget/show-more-text-widget.component';
+import { FoldingCubeComponent } from './components/loaders/folding-cube/folding-cube.component';
+import { LoadingBarsComponent } from './components/loaders/loading-bars/loading-bars.component';
+import { LoadingCirclesComponent } from './components/loaders/loading-circles/loading-circles.component';
+import { RippleCirclesComponent } from './components/loaders/ripple-circles/ripple-circles.component';
+import { VerticalBarsComponent } from './components/loaders/vertical-bars/vertical-bars.component';
+import { WbLoadingListWidgetComponent } from './components/loaders/wb-loading-list-widget/wb-loading-list-widget.component';
 
 @NgModule({
   imports: [
@@ -98,8 +100,8 @@ import { TranslateModule } from '@ngx-translate/core';
     ImageCropperModule,
     InViewportModule
   ],
-  providers: [MenuService, sAuthService, WfLinkPreviewService],
-  declarations: [WfNavbarComponent, WfSideNavComponent, WfNavbarEmptyComponent, WfFooterComponent, WfMainContentComponent, WfLeftSideBarComponent, WfRightSidebarComponent, PostFormComponent, ContentEditableFormDirective, FeedWidgetComponent, LoadingPostWidgetComponent, FeedCommentsWidgetComponent, PostDeleteModalComponent, ReportContentModalComponent, LoadingNewPostWidgetComponent, CommentDeleteModalComponent, LikesModalComponent, LinkPreviewWidgetComponent, RyouMayknowWidgetComponent, ProfileCardWidgetComponent, SearchFormWidgetComponent, MobileSearchFormWidgetComponent, QuickSearchResultsWidgetComponent, SimpleFeedCardWidgetComponent, WelcomePageComponent, WfNavbarSummaryComponent, WfSocialHeaderComponent, WfSocialHeaderComponent, FaintBodyOverlayComponent, FeedCommentRepliesWidgetComponent, CommentReplyDeleteModalComponent, FeedImageViewWidgetComponent, PostImageOverlayViewComponent, FeedPostImageViewComponent, FeedImagesOverlayViewComponent, AddProfilePhotoAlertWidgetComponent, OverlayModalContainerComponent, LoadSubmitProcessDataModalComponent, OverlayLoadSubmitProcessDataWidgetComponent, NavigationWatchComponent, ProcessUploadImagesProgressWidgetComponent, ProfileFeedImageViewWidgetComponent, SharedPostWidgetComponent, OverlayPostViewWidgetComponent, PostCardWidgetComponent, WbMessagesContentComponent, CreatePageModalComponent, ViewAdPostCatalogWidgetComponent, ViewAdPostWidgetComponent, ViewAdTextWidgetComponent, AdTextRightColumnViewComponent, AdTextTopBannerSmallViewComponent, RightSidebarContainerComponent, InPageContentModalComponent, VerticalBarsComponent, LoadingBarsComponent, LoadingCirclesComponent, FoldingCubeComponent, RippleCirclesComponent, MyPagesListWidgetComponent, WbOverlayPostFormContainerComponent, WbPageFollowButtonComponent, WbProfileConnectButtonComponent, WbProfileMessageButtonComponent, WbPageMessageButtonComponent, WbContentNotFoundPageComponent, WbLoadingListWidgetComponent, DeletePostModalComponent, DeleteCommentModalComponent, DeleteConnectionModalComponent],
-  exports: [WfNavbarComponent, WfSideNavComponent, WfNavbarEmptyComponent, WfFooterComponent, WfMainContentComponent, WfLeftSideBarComponent, WfRightSidebarComponent, PostFormComponent, ContentEditableFormDirective, FeedWidgetComponent, LoadingPostWidgetComponent, FeedCommentsWidgetComponent, PostDeleteModalComponent, ReportContentModalComponent, LoadingNewPostWidgetComponent, CommentDeleteModalComponent, LikesModalComponent, LinkPreviewWidgetComponent, RyouMayknowWidgetComponent, ProfileCardWidgetComponent, SearchFormWidgetComponent, MobileSearchFormWidgetComponent, QuickSearchResultsWidgetComponent, SimpleFeedCardWidgetComponent, WelcomePageComponent, WfNavbarSummaryComponent, WfSocialHeaderComponent, FaintBodyOverlayComponent, FeedCommentRepliesWidgetComponent, CommentReplyDeleteModalComponent, FeedImageViewWidgetComponent, PostImageOverlayViewComponent, FeedPostImageViewComponent, FeedImagesOverlayViewComponent, AddProfilePhotoAlertWidgetComponent, OverlayModalContainerComponent, LoadSubmitProcessDataModalComponent, OverlayLoadSubmitProcessDataWidgetComponent, ProcessUploadImagesProgressWidgetComponent, NavigationWatchComponent, ProfileFeedImageViewWidgetComponent, SharedPostWidgetComponent, OverlayPostViewWidgetComponent, PostCardWidgetComponent, WbMessagesContentComponent, CreatePageModalComponent, ViewAdPostCatalogWidgetComponent, ViewAdPostWidgetComponent, ViewAdTextWidgetComponent, AdTextRightColumnViewComponent, AdTextTopBannerSmallViewComponent, RightSidebarContainerComponent, InPageContentModalComponent, VerticalBarsComponent, LoadingBarsComponent, LoadingCirclesComponent, FoldingCubeComponent, RippleCirclesComponent, MyPagesListWidgetComponent, WbOverlayPostFormContainerComponent, WbPageFollowButtonComponent, WbProfileConnectButtonComponent, WbProfileMessageButtonComponent, WbPageMessageButtonComponent, WbContentNotFoundPageComponent, WbLoadingListWidgetComponent,DeletePostModalComponent,DeleteCommentModalComponent, DeleteConnectionModalComponent]
+  providers: [MenuService, WfLinkPreviewService],
+  declarations: [WfNavbarComponent, WfSideNavComponent, WfNavbarEmptyComponent, WfFooterComponent, WfMainContentComponent, WfLeftSideBarComponent, WfRightSidebarComponent, PostFormComponent, ContentEditableFormDirective, FeedWidgetComponent, LoadingPostWidgetComponent, FeedCommentsWidgetComponent, PostDeleteModalComponent, ReportContentModalComponent, LoadingNewPostWidgetComponent, CommentDeleteModalComponent, LikesModalComponent, LinkPreviewWidgetComponent, RyouMayknowWidgetComponent, ProfileCardWidgetComponent, SearchFormWidgetComponent, MobileSearchFormWidgetComponent, QuickSearchResultsWidgetComponent, SimpleFeedCardWidgetComponent, WelcomePageComponent, WfNavbarSummaryComponent, WfSocialHeaderComponent, WfSocialHeaderComponent, FaintBodyOverlayComponent, FeedCommentRepliesWidgetComponent, CommentReplyDeleteModalComponent, FeedImageViewWidgetComponent, PostImageOverlayViewComponent, FeedPostImageViewComponent, FeedImagesOverlayViewComponent, AddProfilePhotoAlertWidgetComponent, OverlayModalContainerComponent, LoadSubmitProcessDataModalComponent, OverlayLoadSubmitProcessDataWidgetComponent, NavigationWatchComponent, ProcessUploadImagesProgressWidgetComponent, ProfileFeedImageViewWidgetComponent, SharedPostWidgetComponent, OverlayPostViewWidgetComponent, PostCardWidgetComponent, WbMessagesContentComponent, CreatePageModalComponent, ViewAdPostCatalogWidgetComponent, ViewAdPostWidgetComponent, ViewAdTextWidgetComponent, AdTextRightColumnViewComponent, AdTextTopBannerSmallViewComponent, RightSidebarContainerComponent, InPageContentModalComponent, VerticalBarsComponent, LoadingBarsComponent, LoadingCirclesComponent, FoldingCubeComponent, RippleCirclesComponent, MyPagesListWidgetComponent, WbOverlayPostFormContainerComponent, WbPageFollowButtonComponent, WbProfileConnectButtonComponent, WbProfileMessageButtonComponent, WbPageMessageButtonComponent, WbContentNotFoundPageComponent, WbLoadingListWidgetComponent, DeletePostModalComponent, DeleteCommentModalComponent, DeleteConnectionModalComponent, ShowMorePostTextWidgetComponent, ShowMoreTextWidgetComponent],
+  exports: [WfNavbarComponent, WfSideNavComponent, WfNavbarEmptyComponent, WfFooterComponent, WfMainContentComponent, WfLeftSideBarComponent, WfRightSidebarComponent, PostFormComponent, ContentEditableFormDirective, FeedWidgetComponent, LoadingPostWidgetComponent, FeedCommentsWidgetComponent, PostDeleteModalComponent, ReportContentModalComponent, LoadingNewPostWidgetComponent, CommentDeleteModalComponent, LikesModalComponent, LinkPreviewWidgetComponent, RyouMayknowWidgetComponent, ProfileCardWidgetComponent, SearchFormWidgetComponent, MobileSearchFormWidgetComponent, QuickSearchResultsWidgetComponent, SimpleFeedCardWidgetComponent, WelcomePageComponent, WfNavbarSummaryComponent, WfSocialHeaderComponent, FaintBodyOverlayComponent, FeedCommentRepliesWidgetComponent, CommentReplyDeleteModalComponent, FeedImageViewWidgetComponent, PostImageOverlayViewComponent, FeedPostImageViewComponent, FeedImagesOverlayViewComponent, AddProfilePhotoAlertWidgetComponent, OverlayModalContainerComponent, LoadSubmitProcessDataModalComponent, OverlayLoadSubmitProcessDataWidgetComponent, ProcessUploadImagesProgressWidgetComponent, NavigationWatchComponent, ProfileFeedImageViewWidgetComponent, SharedPostWidgetComponent, OverlayPostViewWidgetComponent, PostCardWidgetComponent, WbMessagesContentComponent, CreatePageModalComponent, ViewAdPostCatalogWidgetComponent, ViewAdPostWidgetComponent, ViewAdTextWidgetComponent, AdTextRightColumnViewComponent, AdTextTopBannerSmallViewComponent, RightSidebarContainerComponent, InPageContentModalComponent, VerticalBarsComponent, LoadingBarsComponent, LoadingCirclesComponent, FoldingCubeComponent, RippleCirclesComponent, MyPagesListWidgetComponent, WbOverlayPostFormContainerComponent, WbPageFollowButtonComponent, WbProfileConnectButtonComponent, WbProfileMessageButtonComponent, WbPageMessageButtonComponent, WbContentNotFoundPageComponent, WbLoadingListWidgetComponent, DeletePostModalComponent, DeleteCommentModalComponent, DeleteConnectionModalComponent, ShowMorePostTextWidgetComponent, ShowMoreTextWidgetComponent]
 })
 export class ThemeModule { }
